chore: re-organize the anthropic provider
